refactor: move deepCopy out of consts
deepCopy() is not a const, so it should not be in the consts.tsx file.
