+++ app/components/forms/gameState/utils/startTimeHelpers.tsx
@@ -1,6 +1,5 @@
 import { ChangeEvent, Dispatch, SetStateAction } from "react";
-import { deepCopy } from "@/app/utils/consts";
-
+import { deepCopy } from '@/app/utils/utils';
 
 export interface I_StartTimeKit {
     startTime : Date,
