+++ app/utils/consts.tsx
@@ -8,9 +8,7 @@ export const MAX_ALL_EGGS = 5;
 export const TAILWIND_MD_BREAKPOINT = 768;
 export const MS_PER_MINUTE = 60000;
 
-export function deepCopy(target : any){
-    return structuredClone(target);
-}
+
