+++ app/utils/utils.tsx
@@ -1,0 +1,3 @@
+export function deepCopy(target : any){
+    return structuredClone(target);
+}
